FEAT: Make to init logger in __init__.py
