--- app/core/loggertest.py
@@ -1,3 +1,0 @@
-from app.core.test import test
-
-test()
